backend/pyproject.toml
[project]
name = "priotag"
version = "0.1.0"
description = "Small web application to collect images with priorites, analyze them and then generate prio list in an excel sheet"
readme = "README.md"
authors = [
    { name = "DomiDre", email = "dominique.dresen@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "aiofiles>=24.1.0",
    "cryptography>=46.0.2",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "prometheus-client>=0.23.1",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.9",
    "pypdf>=6.0.0",
    "python-dateutil>=2.9.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "types-requests>=2.32.4.20250913",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.1",
    "fakeredis>=2.28.2",
    "ruff>=0.13.0",
    "types-aiofiles>=24.1.0.20250822",
    "types-python-dateutil>=2.9.0.20251108",
    "types-pyyaml>=6.0.12.20250915",
    "types-ujson>=5.10.0.20250822",
    "testcontainers>=4.13.0",
    "redis>=7.0.1",
]
ocr = [
    "matplotlib>=3.9.4",
    "opencv-python>=4.12.0.88",
    "paddleocr[all]>=3.2.0",
    "paddlepaddle>=3.2.0",
]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[tool.setuptools]
packages = [{include = "priotag", from = "src"}]

[tool.ruff]
target-version = "py311"
line-length = 88
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
lint.ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ['pydantic.mypy']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Unit tests that don't require external dependencies",
    "integration: Integration tests that require Redis, PocketBase, etc.",
    "security: Security-focused tests",
    "slow: Tests that take a long time to run",
]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=priotag",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-branch"
]
filterwarnings = [
    "ignore:The @wait_container_is_ready decorator is deprecated:DeprecationWarning:testcontainers",
]
